backend/pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = ""
authors = [
    {name = "suraj"}
]
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
